requirements.txt
amqp==5.2.0
asgiref==3.7.2
billiard==4.2.0
bleach==6.1.0
click==8.1.7
click-didyoumean==0.3.0
click-plugins==1.1.1
click-repl==0.3.0
diff-match-patch==20181111
dj-database-url==1.0.0
Django==4.2.5
django-cors-headers==4.2.0
django-environ==0.11.2
django-filter==23.3
django-quill-editor==0.1.40
django-tinymce==3.6.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
kombu==5.3.4
packaging==23.2
Pillow==10.0.1
prompt-toolkit==3.0.43
protobuf==4.21.12
psycopg2-binary==2.9.5
PyJWT==2.8.0
python-dateutil==2.8.2
python-decouple==3.6
pytz==2023.3.post1
quill-delta==1.0.3
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
vine==5.1.0
waitress==2.1.2
wcwidth==0.2.12
webencodings==0.5.1
whitenoise==6.5.0
